Succesffuly loaded data from online handbook and filled in pre requisites

+++ main.py
@@ -1,7 +1,10 @@
 #from selenium import webdriver
+import os
 import requests
 import json
 import course
+
+degree = {}
 
 def main():
     
@@ -10,65 +13,59 @@ def main():
     specialisationURL = "https://www.handbook.unsw.edu.au/api/content/render/false/query/+contentType:unsw_paos%20+unsw_paos.studyLevelURL:undergraduate%20+unsw_paos.implementationYear:"
     genericURL = "https://www.handbook.unsw.edu.au/api/content/render/false/query/+contentType:unsw_pcourse%20+unsw_pcourse.studyLevelURL:undergraduate%20+unsw_pcourse.implementationYear:"
 
-    courseCode = input("Enter your course code: ")
-
-    
-    implentationYear = input("Enter the year you started: ")
+    degreeCode = input("Enter your course code: ")
+    implementationYear = input("Enter the year you started: ")
 
     courseURL = ''
 
-    if courseCode.isnumeric():
-        courseURL = genericURL + implentationYear + "%20+unsw_pcourse.code:" + courseCode
-    else :
-        courseURL = specialisationURL + implentationYear + "%20+unsw_paos.code:" + courseCode
+    #Check if we have a save for this specific degree and load it up
+
+    degreeFileName = degreeCode+'_'+implementationYear
+
+    if (os.path.isfile('./'+degreeFileName+'.json')):
+        print("loading from save")
+        degreeSave = open(degreeFileName+'.json', 'r')
+        degree = json.load(degreeSave)
+        course.loadCourseListData(degree['courses'])
+        print("Finished loading save")
+
+    else:
+        if degreeCode.isnumeric():
+            courseURL = genericURL + implementationYear + "%20+unsw_pcourse.code:" + degreeCode
+        else :
+            courseURL = specialisationURL + implementationYear + "%20+unsw_paos.code:" + degreeCode
+        
+        
+        degreePage = requests.get(courseURL)
+        print("Searching url: " + courseURL)
     
+        degreeJSON = degreePage.json()['contentlets'][0]
+        curriculumStructure = json.loads(degreeJSON['CurriculumStructure'])
+        noUOC = curriculumStructure['credit_points']
+        f = open("container.txt", "w")
+        f.write(json.dumps(curriculumStructure['container'][0]))
+        f.close()
+
+        dataJSON = json.loads(degreeJSON['data'])
+        coursesJSON = curriculumStructure['container'][0]['container']
+        degreeName = dataJSON['title']
+        degree = {
+            'degreeName' : degreeName,
+            'degreeCode' : degreeCode,
+            'implementationYear': implementationYear,
+            'handbookURL' : courseURL,
+            'courses' :[]
+        }
+
+        print(json.dumps(degree))
+
+
+
+        degree['courses'] = course.getCoursesFromJSON(coursesJSON)
+
+        with open(degreeFileName+'.json', 'w') as json_file:
+            json.dump(degree, json_file)
     
-    degreePage = requests.get(courseURL)
-    print("Searching url: " + courseURL)
-   
-    degreeJSON = degreePage.json()['contentlets'][0]
-
-    curriculumStructure = json.loads(degreeJSON['CurriculumStructure'])
-
-    noUOC = curriculumStructure['credit_points']
-    
-    #print("num uoc = " + noUOC)
-
-    
-    #for key in json.loads(courseJSON['data']).keys():
-    #    print(key)
-
-    
-    f = open("container.txt", "w")
-    f.write(json.dumps(curriculumStructure['container'][0]))
-    f.close()
-
-    dataJSON = json.loads(degreeJSON['data'])
-    courseJSON = curriculumStructure['container'][0]['container']
-    degreeName = dataJSON['title']
-    degree = {
-        'degreeName' : degreeName,
-        'degreeCode' : courseCode,
-        'implementationYear': implentationYear,
-        'handbookURL' : courseURL,
-        'courses' :[]
-    }
-
-    print(json.dumps(degree))
-
-    for degreeLevel in courseJSON:
-        level = degreeLevel['title']
-        order = degreeLevel['order']
-        courses = degreeLevel['relationship']
-        for courseJSON in courses:
-            newCours = {}
-            try :
-                newCourse = course.getCourseData(courseJSON)
-                newCourse['order'] = order
-                newCourse['Level'] = level
-            except:
-                print("Course no longer in use")
-
 
 if __name__ == "__main__":
     main()
